Correccion de implementacion

+++ Punto_Venta/src/main/java/mx/aragon/unam/repository/producto/ProductoRepository.java
@@ -1,9 +1,9 @@
 package mx.aragon.unam.repository.producto;
 
-import mx.aragon.unam.entity.producto.ProductoEntity;
+import mx.aragon.unam.model.entity.producto.ProductoEntity;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface ProductoRepository extends JpaRepository<ProductoEntity, Integer> {
+public interface ProductoRepository extends JpaRepository<ProductoEntity, Long> {
 }
